requirements.txt
<<<<<<< HEAD
Flask==0.10.1
Flask-OAuth==0.12
Jinja2==2.7.1
MarkupSafe==0.18
SOAPpy==0.12.0
Werkzeug==0.9.4
argparse==1.2.1
chardet==2.0.1
fpconst==0.7.2
httplib2==0.8
iotop==0.4.4
itsdangerous==0.23
oauth2==1.5.211
python-oauth2==0.5.0
pymongo==2.2
python-apt==0.8.8.2
python-debian==0.1.21
python-debianbts==1.11
python-firebase==1.1
reportbug==6.4.4
requests==2.1.0
virtualenv==1.7.1.2
wsgiref==0.1.2

=======
Flask==0.8
Flask-OAuth==0.12
Jinja2==2.6
numpy==1.6.2
oauth==1.0.1
oauth2==1.5.211
oauthlib==0.6.1
pymongo==2.6.3
python-firebase==1.1
scikit-learn==0.11
requests==2.2.1
requests-oauthlib==0.4.0
python-digitalocean==0.7
>>>>>>> e833faead1c0ceea92c91566a7cf41a72b5b37b0
